Update WindowOwner.cs
Changes from a direct event call to the generic focus event. The onSetFocusEvent is called within the window via the Focus() functions, but on top of that, the window is also placed up front in the hierarchy.
This prevents windows from being hidden behind the big game window, and leaving players potentially confused.